+++ src/date-range-picker/interfaces.ts
@@ -366,6 +366,8 @@ export namespace DateRangePickerProps {
   }
 }
 
+export type DayIndex = 0 | 1 | 2 | 3 | 4 | 5 | 6;
+
 export interface RangeCalendarValue {
   startDate: string;
   endDate: string;
